+++ DokiApp/APIs/adapters.py
@@ -1,5 +1,3 @@
-from django.shortcuts import render
-
 from ..models import User
 from ..serializers import *
